#53 Moved nordpool api calls and message formatting out of command_sahko to new nordpool_service module
