+++ bobweb/bob/scheduler.py
@@ -13,8 +13,7 @@ from bobweb.bob.utils_common import has
 
 signal.signal(signal.SIGINT, signal.SIG_DFL)
 
-from bobweb.bob import main
-from bobweb.bob import database, broadcaster, command_sahko
+from bobweb.bob import database, broadcaster, nordpool_service
 from bobweb.bob import db_backup
 
 logger = logging.getLogger(__name__)
@@ -51,7 +50,7 @@ class Scheduler:
         # Every midnight empy SahkoCommand cache
         cron_at_midnight_00_00 = '0 0 * * *'
         aiocron.crontab(cron_at_midnight_00_00,
-                        func=command_sahko.cleanup_cache,
+                        func=nordpool_service.cleanup_cache,
                         start=True,
                         tz=fitz)
